backend/requirements.txt
django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
groq==0.9.0
sentence-transformers==2.7.0
chromadb==0.5.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
Pillow==11.0.0
